feature: date 유틸 추가 및 적용 (#81)
